pyproject.toml
[tool.poetry]
name = "wdeditor"
version = "0.1.0"
description = ""
authors = ["Pavel Milanov <pawel.milanov@yandex.ru>"]
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
docxtpl = "^0.16.4"
python-multipart = "^0.0.5"
pydantic = "^1.10.2"
alembic = "^1.9.1"
sqlalchemy = "^1.4.45"
psycopg2-binary = "^2.9.5"
python-jose = "^3.3.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
flake8-builtins = "^2.1.0"
flake8-docstrings = "^1.6.0"
flake8-quotes = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
